examples/image_object_counter_ui/requirements.txt
streamlit
supervision
ultralytics
opencv-python-headless
numpy
pillow
supabase
